refactor: restructure component folders and rename hooks
Reorganize components and rename/move hooks for clarity.
