+++ app/page.tsx
@@ -1,6 +1,6 @@
-import { Banner } from "@/components/page/banner"
-import { SearchBar } from "@/components/page/search-bar"
-import { TrustBadges } from "@/components/page/trust-badges"
+import { Banner } from "@/components/domain/home/banner"
+import { SearchBar } from "@/components/common/search-bar"
+import { TrustBadges } from "@/components/domain/home/trust-badges"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { BookOpen, GraduationCap, Users } from "lucide-react"
