+++ app/search/recruitments/page.tsx
@@ -6,8 +6,8 @@ import { Button } from "@/components/ui/button"
 import { Card, CardContent } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
 import { Tabs, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { RecruitmentCard } from "@/components/page/recruitment-card"
-import { SearchFilters } from "@/components/page/search-filters"
+import { RecruitmentCard } from "@/components/domain/recruitment/recruitment-card"
+import { SearchFilters } from "@/components/domain/recruitment/search-filters"
 
 export default function RecruitmentSearchPage() {
   const [isFilterOpen, setIsFilterOpen] = useState(false)
@@ -74,7 +74,8 @@ export default function RecruitmentSearchPage() {
               professor="김민호 교수"
               deadline="2023-12-31"
               scholarshipAvailable={true}
-              fieldOfStudy="공학"
+              fieldOfStudy="인공지능/머신러닝"
+              location="서울"
             />
             <RecruitmentCard
               lab="경영전략 연구실"
@@ -82,7 +83,8 @@ export default function RecruitmentSearchPage() {
               professor="박지원 교수"
               deadline="2023-11-30"
               scholarshipAvailable={true}
-              fieldOfStudy="경영학"
+              fieldOfStudy="경영학/마케팅"
+              location="서울"
             />
             <RecruitmentCard
               lab="머신러닝 연구실"
@@ -90,7 +92,8 @@ export default function RecruitmentSearchPage() {
               professor="이성훈 교수"
               deadline="2023-12-15"
               scholarshipAvailable={true}
-              fieldOfStudy="공학"
+              fieldOfStudy="컴퓨터 비전/딥러닝"
+              location="대전"
             />
             <RecruitmentCard
               lab="국제관계 연구실"
@@ -98,7 +101,8 @@ export default function RecruitmentSearchPage() {
               professor="최지영 교수"
               deadline="2023-11-15"
               scholarshipAvailable={true}
-              fieldOfStudy="사회과학"
+              fieldOfStudy="국제학/정치학"
+              location="서울"
             />
             <RecruitmentCard
               lab="로봇공학 연구실"
@@ -106,7 +110,8 @@ export default function RecruitmentSearchPage() {
               professor="강현진 교수"
               deadline="2023-12-01"
               scholarshipAvailable={false}
-              fieldOfStudy="공학"
+              fieldOfStudy="로봇공학/제어시스템"
+              location="포항"
             />
             <RecruitmentCard
               lab="바이오메디컬 연구실"
@@ -114,7 +119,8 @@ export default function RecruitmentSearchPage() {
               professor="정민서 교수"
               deadline="2023-12-10"
               scholarshipAvailable={true}
-              fieldOfStudy="생명과학"
+              fieldOfStudy="생명공학/의공학"
+              location="서울"
             />
             <RecruitmentCard
               lab="계량경제학 연구실"
@@ -122,7 +128,8 @@ export default function RecruitmentSearchPage() {
               professor="이지훈 교수"
               deadline="2023-12-20"
               scholarshipAvailable={true}
-              fieldOfStudy="사회과학"
+              fieldOfStudy="경제학/통계학"
+              location="서울"
             />
             <RecruitmentCard
               lab="유기화학 연구실"
@@ -130,7 +137,8 @@ export default function RecruitmentSearchPage() {
               professor="김수민 교수"
               deadline="2023-11-25"
               scholarshipAvailable={false}
-              fieldOfStudy="자연과학"
+              fieldOfStudy="화학/유기합성"
+              location="서울"
             />
             <RecruitmentCard
               lab="나노소재 연구실"
@@ -138,7 +146,8 @@ export default function RecruitmentSearchPage() {
               professor="박혜성 교수"
               deadline="2023-12-05"
               scholarshipAvailable={true}
-              fieldOfStudy="공학"
+              fieldOfStudy="재료공학/나노기술"
+              location="수원"
             />
           </div>
